more refactoring....
