+++ etna/lib/etna/clients/magma/workflows/json_validators.rb
@@ -38,27 +38,6 @@ module Etna
         def model_exists_in_project?(project_magma_models, model_name)
           !!project_magma_models.model(model_name)
         end
-
-        def convert_model_user_json_to_magma_json(model_name, user_json)
-          json_model = user_json.dup
-          json_model['template'] = {
-            'name' => model_name,
-            'attributes' => json_model['attributes'].dup
-          }
-          json_model.delete('attributes')
-          Etna::Clients::Magma::Model.new(json_model)
-        end
-
-        def convert_project_user_json_to_magma_json(user_json)
-          magma_models_json = {}
-          user_json['models'].keys.each do |model_name|
-            magma_models_json[model_name] = convert_model_user_json_to_magma_json(
-              model_name,
-              user_json['models'][model_name])
-          end
-          user_json['models'] = magma_models_json
-          user_json
-        end
       end
 
       class ProjectValidator < ValidatorBase
@@ -66,69 +45,6 @@ module Etna
         def initialize(project)
           super
           @project = project
-          validate
-        end
-
-        def models_by_parent
-          @models_by_parent ||= models.all.group_by { |model| model.raw['parent_model_name'] }
-        end
-
-        def model_tree
-          # Shallow -> deep
-          sorted_models = []
-          parent_model_names = [nil]
-
-          loop do
-            child_models = models_by_parent.values_at(*parent_model_names).flatten.compact
-
-            break if child_models.length == 0
-
-            parent_model_names = child_models.map { |model| model.name }
-
-            # Sort by name within each depth level
-            sorted_models += child_models.sort do |m1, m2|
-              m1.name <=> m2.name
-            end
-          end
-          sorted_models
-        end
-
-        def update_magma_models
-          models.each do |model|
-            # Because the input JSON format doesn't specify the child
-            #   or reciprocal link attributes, we'll need to add those
-            #   in manually.
-            add_child_attributes(model_builder, model)
-          end
-
-          add_reciprocal_link_attributes(magma_models)
-        end
-
-        def add_child_attributes(model)
-          return unless models_by_parent[model.name]
-
-          models_by_parent[model.name].each do |child_model|
-            attribute_builder = model.build_template.build_attributes
-            add_child_attribute(attribute_builder, child_model)
-          end
-        end
-
-        def add_child_attribute(builder, child_model)
-          builder.build_attribute(child_model.name).tap do |attribute|
-            attribute.attribute_name = child_model.name
-            attribute.name = child_model.name
-            attribute.attribute_type = child_model.parent_link_type
-            attribute.display_name = child_model.prettified_name
-            attribute.desc = child_model.prettified_name
-          end
-        end
-
-        def add_reciprocal_link_attributes(magma_models)
-          all_link_attributes do |model, attribute|
-            reciprocal_model = project.models.model(attribute.link_model_name)
-            model_builder = magma_models.build_model(reciprocal_model.name)
-            reciprocal_model.add_reciprocal_link_attribute(model_builder, model)
-          end
         end
 
         def validate
@@ -146,8 +62,8 @@ module Etna
 
         def validate_models
           project.models.all.each do |model|
-            validator = ModelValidator.new
-            validator.validate_model(model)
+            validator = ModelValidator.new(model)
+            validator.validate_model
             @errors += validator.errors unless validator.valid?
           end
         end
@@ -157,95 +73,35 @@ module Etna
           #   that the link_model_name exists in the project definition.
           model_names = project.models.all.map { |m| m.name }
           project.models.all.map do |model|
-            validator = ModelValidator.new
+            validator = ModelValidator.new(model)
             validator.validate_link_models(model_names)
             @errors += validator.errors unless validator.valid?
-          end
-        end
-
-        def all_link_attributes
-          models.map do |model|
-            model.link_attributes do |attribute|
-              yield [model, attribute]
-            end
           end
         end
       end
 
       class ModelValidator < ValidatorBase
-        attr_reader :name
-        def initialize(model_name, raw)
+        attr_reader :model
+        def initialize(model)
           super()
-          @name = model_name.strip
-          @raw = raw
-
+          @model = model
           @valid_parent_link_types = Etna::Clients::Magma::ParentLinkType.entries.sort # sort for prettier presentation
-
-          validate
         end
 
-        def self.linking_stub_from_name(model_name)
-          # Used for linking to a link_model_name only! Fills out required fields with
-          #   dummy information to pass validation.
-          Etna::Clients::Magma::JsonModel.new(
-            model_name,
-            {
-              "parent_model_name" => "stub",
-              "parent_link_type" => "child",
-              "identifier" => "none",
-              "attributes" => {}
-            }
-          )
+        def name
+          model.name
+        end
+
+        def raw
+          model.raw
         end
 
         def parent_model_name
-          @raw['parent_model_name']&.strip
-        end
-
-        def prettified_parent_model_name
-          parent_model_name.split('_').map(&:capitalize).join(' ')
-        end
-
-        def parent_link_type
-          @raw['parent_link_type']&.strip
-        end
-
-        def identifier
-          @raw['identifier']&.strip
-        end
-
-        def prettified_name
-          name.split('_').map(&:capitalize).join(' ')
+          model.raw['parent_model_name']
         end
 
         def is_project?
-          name == 'project'
-        end
-
-        def attributes
-          @attributes ||= @raw.key?('attributes') ? @raw['attributes'].map do |attribute_name, attribute_def|
-            attribute = Etna::Clients::Magma::Attribute.new(attribute_def)
-            attribute.attribute_type = Etna::Clients::Magma::AttributeType::IDENTIFIER if identifier == attribute.attribute_name
-            attribute
-          end: []
-        end
-
-        def link_attributes
-          attributes.map do |attribute|
-            yield attribute if attribute.is_link_attribute?
-          end
-        end
-
-        def add_reciprocal_link_attribute(builder, model)
-          attribute_builder = builder.build_template.build_attributes
-          attribute_builder.build_attribute(model.name).tap do |attribute|
-            attribute.attribute_name = model.name
-            attribute.name = model.name
-            attribute.attribute_type = Etna::Clients::Magma::AttributeType::COLLECTION
-            attribute.display_name = model.prettified_name
-            attribute.desc = model.prettified_name
-            attribute.link_model_name = model.name
-          end
+          model.name == 'project'
         end
 
         def validate
@@ -254,32 +110,33 @@ module Etna
         end
 
         def validate_add_model_data
-          @errors << "Model name #{name} must be snake_case and can only consist of letters and \"_\"." unless name =~ name_regex_no_numbers
+          @errors << "Model name #{name} must be snake_case and can only consist of letters and \"_\"." unless model.name =~ name_regex_no_numbers
 
           if !is_project?
-            check_key("model #{name}", @raw, 'parent_model_name')
-            check_key("model #{name}", @raw, 'parent_link_type')
-            check_type("model #{name}", @raw, 'parent_link_type', @valid_parent_link_types)
+            check_key("model #{name}", raw, 'parent_model_name')
+            check_key("model #{name}", raw, 'parent_link_type')
+            check_type("model #{name}", raw, 'parent_link_type', @valid_parent_link_types)
           end
-          if parent_link_type != Etna::Clients::Magma::ParentLinkType::TABLE
-            check_key("model #{name}", @raw, 'identifier')
+          if model.raw['parent_link_type'] != Etna::Clients::Magma::ParentLinkType::TABLE
+            check_key("model #{name}", raw, 'identifier')
           end
         end
 
         def validate_attributes
-          attributes.each do |attribute|
-            attribute_validator = AttributeValidator.new
-            attribute_validator.validate_add_attribute_data(attribute)
+          attributes.all.each do |attribute|
+            attribute_validator = AttributeValidator.new(attribute)
+            attribute_validator.validate_add_attribute_data
             @errors += attribute_validator.errors unless attribute_validator.valid?
           end
         end
 
         def validate_link_models(model_names)
-          @errors << "Parent model #{parent_model_name} for #{name} does not exist in project.\nCurrent models are #{model_names}." if name != 'project' && !model_names.include?(parent_model_name)
+          @errors << "Parent model #{parent_model_name} for #{name} does not exist in project.\nCurrent models are #{model_names}." if !is_project? && !model_names.include?(parent_model_name)
 
           link_attributes do |attribute|
-            attribute.validate_link_models(model_names)
-            @errors = (@errors + attribute.errors).uniq # need to re-add attribute.errors because validate_link_models called at project level.
+            attribute_validator = AttributeValidator.new(attribute)
+            attribute_validator.validate_link_models(model_names)
+            @errors += attribute_validator.errors unless attribute_validator.valid?
 
             @errors << "Model \"#{name}\" already belongs to parent model \"#{parent_model_name}\". Remove attribute \"#{attribute.name}\"." if attribute.link_model_name == parent_model_name
           end
@@ -288,37 +145,13 @@ module Etna
         def validate_existing_models(project_magma_models)
           @errors << "Model #{name} already exists in project!" if model_exists_in_project?(project_magma_models, name)
         end
-
-        def to_magma_model(builder)
-          template_builder = builder.build_template
-          template_builder.identifier = identifier
-          template_builder.parent = parent_model_name
-          attribute_builder = template_builder.build_attributes
-          attributes.each do |attribute|
-            attribute.to_magma_model(attribute_builder)
-          end
-
-          # Because the input JSON format doesn't specify the parent
-          #   attribute, we'll need to add those
-          #   in manually.
-          add_parent_attribute(attribute_builder) unless is_project?
-        end
-
-        def add_parent_attribute(builder)
-          builder.build_attribute(parent_model_name).tap do |attribute|
-            attribute.attribute_name = parent_model_name
-            attribute.name = parent_model_name
-            attribute.link_model_name = parent_model_name
-            attribute.attribute_type = Etna::Clients::Magma::AttributeType::PARENT
-            attribute.display_name = prettified_parent_model_name
-            attribute.desc = prettified_parent_model_name
-          end
-        end
       end
 
       class AttributeValidator < ValidatorBase
-        def initialize
+        attr_reader :attribute
+        def initialize(attribute)
           super
+          @attribute = attribute
           # NOTE: for input simplicity, I've removed some of the
           #   link-related types, since we'll try to calculate
           #   those while parsing the JSON structure.
@@ -336,7 +169,7 @@ module Etna
           @errors << "#{label} name #{proposed_name} must be snake_case and can only consist of letters, numbers, and \"_\"." unless proposed_name =~ name_regex_with_numbers
         end
 
-        def validate_basic_attribute_data(attribute)
+        def validate_basic_attribute_data
           check_key("attribute #{attribute.attribute_name}", attribute.raw, 'attribute_type')
 
           # The following two could be calculated or left blank?
@@ -477,7 +310,7 @@ module Etna
       #   end
       # end
 
-      class JsonAttributeAction < ValidatorBase
+      class JsonAttributeActionValidator < ValidatorBase
         def initialize(raw, project_models)
           @raw = raw
           @project_models = project_models
@@ -509,7 +342,7 @@ module Etna
         end
       end
 
-      class JsonAddAttributeAction < JsonAttributeAction
+      class JsonAddAttributeActionValidator < JsonAttributeActionValidator
         def initialize
           super
           @attribute = Etna::Clients::Magma::Attribute.new(@raw)
@@ -528,7 +361,7 @@ module Etna
         end
       end
 
-      class JsonAddLinkAction < JsonAttributeAction
+      class JsonAddLinkActionValidator < JsonAttributeActionValidator
         def links
           @raw['links']
         end
@@ -578,7 +411,7 @@ module Etna
         end
       end
 
-      class JsonRenameAttributeAction < JsonAttributeAction
+      class JsonRenameAttributeActionValidator < JsonAttributeActionValidator
         def new_attribute_name
           @raw['new_attribute_name']&.strip
         end
@@ -601,7 +434,7 @@ module Etna
         end
       end
 
-      class JsonUpdateAttributeAction < JsonAttributeAction
+      class JsonUpdateAttributeActionValidator < JsonAttributeActionValidator
         def initialize
           super
           @attribute = Etna::Clients::Magma::Attribute.new(@raw)
@@ -619,13 +452,13 @@ module Etna
         end
       end
 
-      class JsonAttributeActionFactory
+      class JsonAttributeActionValidatorValidatorFactory
         def camelize(action_name)
           action_name.split('_').map(&:capitalize).join('')
         end
 
         def clazz_name(action_name)
-          "Json#{camelize(action_name)}Action"
+          "Json#{camelize(action_name)}ActionValidator"
         end
 
         def self.from_json(raw, project_models)
